+++ src/main/java/org/craftercms/studio/impl/v1/service/deployment/job/DeployContentToEnvironmentStore.java
@@ -53,6 +53,7 @@ import org.craftercms.studio.api.v1.to.PublishingTargetTO;
 import org.craftercms.studio.api.v1.util.StudioConfiguration;
 import org.craftercms.studio.api.v2.service.notification.NotificationService;
 import org.craftercms.studio.impl.v1.job.RepositoryJob;
+import org.springframework.core.task.TaskExecutor;
 
 import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOYMENT_MASTER_PUBLISHING_NODE;
 import static org.craftercms.studio.api.v1.util.StudioConfiguration.JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_MANDATORY_DEPENDENCIES_CHECK_ENABLED;
@@ -82,6 +83,7 @@ public class DeployContentToEnvironmentStore extends RepositoryJob {
     protected StudioConfiguration studioConfiguration;
     protected ActivityService activityService;
     protected ServicesConfig servicesConfig;
+    protected TaskExecutor taskExecutor;
 
     public static synchronized void signalToStop(boolean toStop) {
         stopSignaled = toStop;
@@ -138,87 +140,10 @@ public class DeployContentToEnvironmentStore extends RepositoryJob {
                                         logger.debug("Site \"{0}\" has {1} items ready for deployment",
                                                 site, itemsToDeploy.size());
 
-                                        String author = itemsToDeploy.get(0).getUser();
-                                        StringBuilder sbComment = new StringBuilder();
-                                        List<DeploymentItemTO> completeDeploymentItemList = new ArrayList<DeploymentItemTO>();
-                                        Set<String> processedPaths = new HashSet<String>();
-                                        SimpleDateFormat sdf =
-                                                new SimpleDateFormat(StudioConstants.DATE_PATTERN_WORKFLOW_WITH_TZ);
-                                        String messagePath = StringUtils.EMPTY;
-                                        String currentPackageId = StringUtils.EMPTY;
-                                        try {
-                                            logger.debug("Mark items as processing for site \"{0}\"", site);
-                                            Set<String> packageIds = new HashSet<String>();
-                                            for (PublishRequest item : itemsToDeploy) {
-                                                processPublishingRequest(site, environment, item,
-                                                        completeDeploymentItemList, processedPaths);
-                                                if (!StringUtils.equals(currentPackageId, item.getPackageId())) {
-                                                    currentPackageId = item.getPackageId();
-                                                    statusMessage = studioConfiguration.getProperty
-                                                            (JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_PUBLISHING);
-                                                    statusMessage =
-                                                            statusMessage.replace("{package_id}", currentPackageId)
-                                                            .replace("{datetime}", ZonedDateTime.now(ZoneOffset.UTC)
-                                                                    .format(DateTimeFormatter.ofPattern(sdf.toPattern())));
-                                                    siteService.updatePublishingStatusMessage(site, statusMessage);
-                                                }
-                                                if (packageIds.add(item.getPackageId())) {
-                                                    sbComment.append(item.getSubmissionComment()).append("\n");
-                                                }
-                                            }
-                                            deploy(site, environment, completeDeploymentItemList, author,
-                                                    sbComment.toString());
-                                            StringBuilder sbPackIds = new StringBuilder("Package(s): ");
-                                            for (String packageId : packageIds) {
-                                                sbPackIds.append(packageId).append(";");
-                                            }
-                                            generateWorkflowActivity(site, sbPackIds.toString(), author,
-                                                    ActivityService .ActivityType .PUBLISHED);
-                                            publishingManager.markItemsCompleted(site, environment, itemsToDeploy);
-                                            logger.debug("Mark deployment completed for processed items for site \"{0}\"", site);
-                                            logger.info("Finished publishing environment " + environment + " for site " + site);
-
-                                            if (publishingManager.isPublishingQueueEmpty(site)) {
-                                                statusMessage = studioConfiguration.getProperty
-                                                        (JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_IDLE);
-                                                statusMessage = statusMessage.replace("{package_id}", currentPackageId)
-                                                        .replace("{datetime}", ZonedDateTime.now(ZoneOffset.UTC)
-                                                                .format(DateTimeFormatter.ofPattern(sdf.toPattern())))
-                                                        .replace("{package_size}",
-                                                                Integer.toString(itemsToDeploy.size ()));
-                                            } else {
-                                                statusMessage =
-                                                        studioConfiguration.getProperty
-                                                                (JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_QUEUED);
-                                            }
-                                            siteService.updatePublishingStatusMessage(site, statusMessage);
-                                        } catch (DeploymentException err) {
-                                            logger.error("Error while executing deployment to environment store " +
-                                                    "for site \"{0}\", number of items \"{1}\"", err, site,
-                                                    itemsToDeploy.size());
-                                            publishingManager.markItemsReady(site, environment, itemsToDeploy);
-                                            siteService.enablePublishing(site, false);
-                                            statusMessage = studioConfiguration.getProperty
-                                                    (JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_STOPPED_ERROR);
-                                            statusMessage = statusMessage.replace("{item_path}", messagePath)
-                                                    .replace("{datetime}", ZonedDateTime.now(ZoneOffset.UTC)
-                                                            .format(DateTimeFormatter.ofPattern(sdf.toPattern())));
-                                            siteService.updatePublishingStatusMessage(site, statusMessage);
-                                            throw err;
-                                        } catch (Exception err){
-                                            logger.error("Unexpected error while executing deployment to environment " +
-                                                    "store for site \"{0}\", number of items \"{1}\"", err, site,
-                                                    itemsToDeploy.size());
-                                            publishingManager.markItemsReady(site, environment, itemsToDeploy);
-                                            siteService.enablePublishing(site, false);
-                                            statusMessage = studioConfiguration.getProperty
-                                                    (JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_STOPPED_ERROR);
-                                            statusMessage = statusMessage.replace("{item_path}", messagePath)
-                                                    .replace("{datetime}", ZonedDateTime.now(ZoneOffset.UTC)
-                                                            .format(DateTimeFormatter.ofPattern(sdf.toPattern())));
-                                            siteService.updatePublishingStatusMessage(site, statusMessage);
-                                            throw err;
-                                        }
+                                        PublisherTask publisherTask = new PublisherTask(site, itemsToDeploy,
+                                                environment, studioConfiguration, siteService, publishingManager,
+                                                servicesConfig,contentRepository, activityService, notificationService);
+                                        taskExecutor.execute(publisherTask);
                                     }
                                 }
                             } catch (Exception err) {
@@ -239,89 +164,6 @@ public class DeployContentToEnvironmentStore extends RepositoryJob {
             logger.error("Error while executing deployment to environment store", err);
             notificationService.notifyDeploymentError("UNKNOWN", err);
         }
-    }
-
-    protected void generateWorkflowActivity(String site, String path, String username,
-                                            ActivityService.ActivityType activityType) {
-        Map<String, String> extraInfo = new HashMap<String, String>();
-        extraInfo.put(DmConstants.KEY_CONTENT_TYPE, StudioConstants.CONTENT_TYPE_USER);
-        activityService.postActivity(site, username, path, activityType,
-                ActivityService.ActivitySource.API, extraInfo);
-    }
-
-    private void processPublishingRequest(String site, String environment, PublishRequest item,
-                                          List<DeploymentItemTO> completeDeploymentItemList,
-                                          Set<String> processedPaths)
-        throws ServiceLayerException, DeploymentException {
-        List<DeploymentItemTO> missingDependencies = new ArrayList<DeploymentItemTO>();
-        Set<String> missingDependenciesPaths = new HashSet<String>();
-        SimpleDateFormat sdf = new SimpleDateFormat(StudioConstants.DATE_PATTERN_WORKFLOW_WITH_TZ);
-        String messagePath = item.getPath();
-        String statusMessage = studioConfiguration.getProperty(JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_BUSY);
-        statusMessage = statusMessage.replace("{item_path}", messagePath).replace("{datetime}",
-                ZonedDateTime.now(ZoneOffset.UTC).format(DateTimeFormatter.ofPattern(sdf.toPattern())));
-        siteService.updatePublishingStatusMessage(site, statusMessage);
-        publishingManager.markItemsProcessing(site, environment, Arrays.asList(item));
-        try {
-            List<DeploymentItemTO> deploymentItemList = new ArrayList<DeploymentItemTO>();
-
-
-            logger.debug("Processing [{0}] content item for site \"{1}\"", item
-                    .getPath(), site);
-            DeploymentItemTO deploymentItem = publishingManager.processItem(item);
-            if (deploymentItem != null) {
-                deploymentItemList.add(deploymentItem);
-            }
-            logger.debug("Processing COMPLETE [{0}] content item for site \"{1}\"",
-                    item.getPath(), site);
-
-            if (isMandatoryDependenciesCheckEnabled()) {
-                logger.debug("Processing Mandatory Deps [{0}] content item for site "
-                        + "\"{1}\"", item.getPath(), site);
-                missingDependencies.addAll(publishingManager
-                        .processMandatoryDependencies(item, processedPaths, missingDependenciesPaths));
-                logger.debug("Processing Mandatory Dependencies COMPLETE [{0}]"
-                        + " content item for site \"{1}\"", item.getPath(), site);
-            }
-            deploymentItemList.addAll(missingDependencies);
-            completeDeploymentItemList.addAll(deploymentItemList);
-        } catch (DeploymentException err) {
-            logger.error("Error while executing deployment to environment store for site \"{0}\",", err, site);
-            publishingManager.markItemsReady(site, environment, Arrays.asList(item));
-            siteService.enablePublishing(site, false);
-            statusMessage = studioConfiguration.getProperty(
-                    JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_STOPPED_ERROR);
-            statusMessage = statusMessage.replace("{item_path}", messagePath).replace("{datetime}",
-                    ZonedDateTime.now(ZoneOffset.UTC).format(DateTimeFormatter.ofPattern(sdf.toPattern())));
-            siteService.updatePublishingStatusMessage(site, statusMessage);
-            throw err;
-        } catch (Exception err){
-            logger.error("Unexpected error while executing deployment to environment " +
-                    "store for site \"{0}\", ", err, site);
-            publishingManager.markItemsReady(site, environment, Arrays.asList(item));
-            siteService.enablePublishing(site, false);
-            statusMessage = studioConfiguration.getProperty(
-                    JOB_DEPLOY_CONTENT_TO_ENVIRONMENT_STATUS_MESSAGE_STOPPED_ERROR);
-            statusMessage = statusMessage.replace("{item_path}", messagePath).replace("{datetime}",
-                    ZonedDateTime.now(ZoneOffset.UTC).format(DateTimeFormatter.ofPattern(sdf.toPattern())));
-            siteService.updatePublishingStatusMessage(site, statusMessage);
-            throw err;
-        }
-    }
-
-    private void deploy(String site, String environment, List<DeploymentItemTO> items, String author, String comment)
-            throws DeploymentException, SiteNotFoundException {
-        logger.debug("Deploying " + items.size() + " item(s)");
-        SiteFeed siteFeed = siteService.getSite(site);
-        if (servicesConfig.isStagingEnvironmentEnabled(site)) {
-            String liveEnvironment = servicesConfig.getLiveEnvironment(site);
-            if (StringUtils.equals(liveEnvironment, environment)) {
-                String stagingEnvironment = servicesConfig.getStagingEnvironment(site);
-                contentRepository.publish(site, siteFeed.getSandboxBranch(), items, stagingEnvironment, author, comment);
-            }
-        }
-        contentRepository.publish(site, siteFeed.getSandboxBranch(), items, environment, author, comment);
-
     }
 
     private Set<String> getAllPublishingEnvironments(String site) {
@@ -441,5 +283,13 @@ public class DeployContentToEnvironmentStore extends RepositoryJob {
 
     public void setServicesConfig(ServicesConfig servicesConfig) {
         this.servicesConfig = servicesConfig;
+    }
+
+    public TaskExecutor getTaskExecutor() {
+        return taskExecutor;
+    }
+
+    public void setTaskExecutor(TaskExecutor taskExecutor) {
+        this.taskExecutor = taskExecutor;
     }
 }
